Add Installation Message Dev Override

+++ admin/pireps.php
@@ -38,7 +38,7 @@ $ACTIVE_CATEGORY = 'pirep-management';
                     <div id="loader-wrapper"><div id="loader" class="spinner-border spinner-border-sm spinner-custom"></div></div>
                     <div class="loaded">
                         <?php
-                        if (file_exists(__DIR__.'/../install/install.php')) {
+                        if (file_exists(__DIR__.'/../install/install.php') && !file_exists(__DIR__.'/../.development')) {
                             echo '<div class="alert alert-danger text-center">The Install Folder still Exists! Please delete it immediately, it poses a severe security risk.</div>';
                         }
